+++ fm/src/app/app-routing.module.ts
@@ -56,6 +56,7 @@ import { BirthdayCardComponent } from './birthday-card/birthday-card.component';
 import { OperationComponent } from './operation/operation.component';
 import { ReportsComponent } from './reports/reports.component';
 import { NewJoinersComponent } from './new-joiners/new-joiners.component';
+import { FeedbackComponent } from './feedback/feedback.component';
 
 const routes: Routes = [
   { path: 'general', component: GeneralComponent },
@@ -77,6 +78,7 @@ const routes: Routes = [
   {path:'reports',component:ReportsComponent},
   { path: 'newHires', component: NewJoinersComponent },
   {path:'details', component:DetailsComponent},
+  {path:'feedback', component:FeedbackComponent},
 
   {
     path: 'leave-tracking',
@@ -90,6 +92,8 @@ const routes: Routes = [
           { path: 'leave-balance', component: LeaveBalanceComponent },
           { path: 'leave-requests', component: LeaveRequestsComponent },
           { path: 'apply-leave', component: ApplyLeaveComponent },
+          { path: 'feedback', component: FeedbackComponent },
+
         ],
       },
       {
