+++ fm/src/app/app.module.ts
@@ -91,11 +91,13 @@ import { FeedsComponent } from './feeds/feeds.component';
 import { OperationComponent } from './operation/operation.component';
 import { ReportsComponent } from './reports/reports.component';
 import { NewJoinersComponent } from './new-joiners/new-joiners.component';
+import { FeedbackComponent } from "./feedback/feedback.component";
 
 @NgModule({
   declarations: [
     NewJoinersComponent,
     AppComponent,
+    FeedbackComponent,
     GeneralComponent,
     EducationalComponent,
     SkillsetComponent,
@@ -187,8 +189,8 @@ import { NewJoinersComponent } from './new-joiners/new-joiners.component';
     MatStepperModule,
     NgIf,
     NgFor,
-    TreeModule,
-  ],
+    TreeModule
+],
   providers: [],
   bootstrap: [AppComponent],
 })
