+++ fm/src/app/feedback/feedback.component.scss
@@ -1,0 +1,91 @@
+.feedback-container {
+    margin: 20px;
+  }
+  
+  .feedback-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+  
+  .feedback-header button {
+    padding: 10px 20px;
+    font-size: 16px;
+    background-color: #f0f0f0;
+    border: none;
+    cursor: pointer;
+    outline: none;
+  }
+  
+  .feedback-header button.active {
+    background-color: #0056b3;
+    color: white;
+  }
+  
+  .search-input {
+    padding: 10px;
+    font-size: 16px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    width: 250px;
+  }
+  
+  .feedback-filters {
+    display: flex;
+    justify-content: space-around;
+    margin-top: 20px;
+  }
+  
+  .feedback-card {
+    width: 120px;
+    height: 100px;
+    border-radius: 8px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+  }
+  
+  .feedback-card h4 {
+    margin: 0;
+    font-size: 16px;
+    color: #333;
+  }
+  
+  .feedback-card p {
+    margin: 5px 0 0 0;
+    font-size: 24px;
+  }
+  
+  .feedback-card.all {
+    background-color: #e0f7fa;
+  }
+  
+  .feedback-card.positive {
+    background-color: #c8e6c9;
+  }
+  
+  .feedback-card.negative {
+    background-color: #ffcdd2;
+  }
+  
+  .feedback-card.training {
+    background-color: #ffecb3;
+  }
+  
+  .feedback-card.observation {
+    background-color: #bbdefb;
+  }
+  
+  .feedback-card.reward {
+    background-color: #ffe0b2;
+  }
+  
+  .feedback-content {
+    margin-top: 40px;
+    text-align: center;
+    font-size: 18px;
+  }
+
