+++ fm/src/app/overview/overview.component.ts
@@ -1,5 +1,5 @@
 import { Component, ElementRef, ViewChild } from '@angular/core';
-import * as bootstrap from 'bootstrap';
+// import * as bootstrap from 'bootstrap';
 
 @Component({
   selector: 'app-overview',
